+++ src/engine/resolver.spec.ts
@@ -55,6 +55,7 @@ describe('Resolver', () => {
 			findDefs: jest.fn(),
 			findDefsByIds: jest.fn(),
 			findAnyDef: jest.fn(),
+			findAnyDefs: jest.fn(),
 			findValues: jest.fn(),
 			findChainValues: jest.fn(),
 			findAudit: jest.fn(),